% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itan.R
\name{calcularPuntajes}
\alias{calcularPuntajes}
\title{C\enc{á}{a}lculo de puntajes.}
\usage{
calcularPuntajes(respCorregidas)
}
\arguments{
\item{respCorregidas}{Una matriz con las respuestas de los estudiantes a cada \enc{í}{i}tem
en forma binaria.}
}
\value{
Un vector con los puntajes obtenidos por los estudiantes.
}
\description{
Calcula el puntaje total obtenido en la prueba por cada estudiante.
}
\examples{
data(datos)
data(clave)
respuestas <- subset(datos, select = -id )
puntajes <- calcularPuntajes(corregirItems(respuestas, clave))
cbind(respuestas, pje=puntajes)

}
\seealso{
\code{\link{calcularNotas}} y \code{\link{corregirItems}}.
}
