% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clave}
\alias{clave}
\title{Respuestas correctas a los \enc{í}{i}tems del test}
\format{
Data frame con 1 observación y 50 variables:
\describe{
\item{i01}{Alternativa correcta para la pregunta 1.}
\item{i02}{Alternativa correcta para la pregunta 2.}
...
\item{i50}{Alternativa correcta para la pregunta 50.}
}
}
\usage{
clave
}
\description{
Un data frame con las respuestas correctas a los \enc{í}{i}tems del test.
}
\keyword{datasets}
