% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itan.R
\name{corregirItems}
\alias{corregirItems}
\title{Correci\enc{ó}{o}n de \enc{í}{i}tems.}
\usage{
corregirItems(respuestas, clave)
}
\arguments{
\item{respuestas}{Una matriz con las alternativas seleccionadas por los
estudiantes a cada \enc{í}{i}tem}

\item{clave}{Una matriz con la alternativa correcta para cada \enc{í}{i}tem.}
}
\value{
Una matriz binaria con los aciertos (1) o errores (0) de
cada estudianta en cada \enc{í}{i}tem.
}
\description{
Transforma una matriz con las alternativas seleccionadas por
los estudiantes en cada \enc{í}{i}tem en una matriz binaria (correcto/incorrecto) segun
la clave de correcci\enc{ó}{o}n.
}
\examples{
data(datos)
data(clave)
respuestas <- subset(datos, select = -id )
corregirItems(respuestas, clave)

}
\seealso{
\code{\link{analizarItems}}
}
