% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itan.R
\name{corregirRespuestas}
\alias{corregirRespuestas}
\title{Corrección de respuestas}
\usage{
corregirRespuestas(respuestas, clave)
}
\arguments{
\item{respuestas}{Un data frame con las alternativas seleccionadas por los
estudiantes en cada ítem.}

\item{clave}{Una data frame con la alternativa correcta para cada ítem.}
}
\value{
Un data frame con los aciertos (1) o errores (0) de
cada estudiante en cada ítem.
}
\description{
La función corregir respuestas transforma las respuestas de los estudiantes
a puntaje. El puntaje puede ser un \code{1}, si la respuesta es correcta, o un
\code{0}, si la respuesta es incorrecta. Los valores \code{NA} reciben puntaje
\code{0}.
}
\examples{
respuestas <- datos[, -1]
corregirRespuestas(respuestas, clave)

}
\seealso{
\code{\link{datos}} y \code{\link{clave}}
}
