% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datos}
\alias{datos}
\title{Datos de los estudiantes}
\format{
Data frame con 39 observaciones y 51 variables:
\describe{
\item{id}{Id del estudiante.}
\item{i01}{Alternativa seleccionada por el estudiante a la pregunta 1.}
\item{i02}{Alternativa seleccionada por el estudiante a la pregunta 2.}
...
\item{i50}{Alternativa seleccionada por el estudiante a la pregunta 50.}
}
}
\usage{
datos
}
\description{
Un data frame con el id  y las respuestas de los estudiantes.
}
\keyword{datasets}
