% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itan.R
\name{frecuenciaDistractores}
\alias{frecuenciaDistractores}
\title{Frecuencia de distractores.}
\usage{
frecuenciaDistractores(respuestas, clave = NULL, frecuencia = TRUE,
  digitos = 2)
}
\arguments{
\item{respuestas}{Una hoja de datos con las alternativas que seleccion\enc{ó}{o} cada
estudiante en cada \enc{í}{i}tem de la prueba.}

\item{clave}{Una hoja de datos con la alternativa correcta para cada \enc{í}{i}tem.}

\item{frecuencia}{Un valor l\enc{ó}{o}gico que determina si la informacin\enc{ó}{o}
se presenta como frecuencia (TRUE) o proporci\enc{ó}{o}n (FALSE).}

\item{digitos}{La cantidad de d\enc{í}{i}gitos significativos que tendr\enc{á}{a} el resultado.}
}
\value{
Una matriz con los \enc{í}{i}tems como filas y las frecuencias de las
alternativas como columnas. Si se entrega como argumento la clave de la prueba, se entrega una
hoja de datos con la frecuencia o proporci\enc{ó}{o}n de las alternativas y la alternativa
correcta para cada \enc{í}{i}tem.
}
\description{
Calcula la frecuencia o proporci\enc{ó}{o}n de cada alternativa seleccionada
en cada \enc{í}{i}tem.
}
\examples{
data(datos)
respuestas <- subset(datos, select = -id)
frecuenciaDistractores(respuestas)

}
\seealso{
\code{\link{analizarDistractores}}.
}
