% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itan.R
\name{calcularPuntajes}
\alias{calcularPuntajes}
\title{Cálculo de puntajes}
\usage{
calcularPuntajes(respuestas)
}
\arguments{
\item{respuestas}{Un data frame con el puntaje obtenido por los estudiantes
en cada ítem.}
}
\value{
Un vector con el puntaje total obtenido en la pruebas por cada estudiante.
}
\description{
Calcula el puntaje total obtenido en la prueba por cada estudiante.
}
\examples{
respuestas <- datos[,-1]
respuestasCorregidas <- corregirRespuestas(respuestas, clave)
puntajes <- calcularPuntajes(respuestasCorregidas)
cbind(id=datos[1], puntaje=puntajes)

}
\seealso{
\code{\link{corregirRespuestas}}, \code{\link{datos}} y \code{\link{clave}}.
}
